I-#204: Refactor notificaiton.
