--- clients/mobile/src/setup/notification/NotificationContext.tsx
@@ -1,5 +1,0 @@
-import { NotificationData } from './NotificationData'
-
-export type NotificationContextType = {
-  notificationData: NotificationData | null;
-}
