+++ clients/mobile/src/setup/notification/NotificationData.ts
@@ -1,9 +1,5 @@
 export type NotificationData = Record<string, unknown> | null
 
-export type NotificationDataTypes = {
-  notificationData: Record<string, unknown> | null;
-}
-
-export const initialNotificationState: NotificationDataTypes = {
-  notificationData: null
+export interface NotificationDataTypes {
+  notificationData: NotificationData;
 }
