+++ clients/mobile/src/setup/notification/saveDeviceToken.ts
@@ -1,9 +1,7 @@
 import { Platform } from 'react-native'
+import { HttpClient } from '../clients/HttpClient'
 
-export const saveDeviceTokenToSNS = async(
-  deviceToken: string,
-  fetchClient: any
-): Promise<void> => {
+export const saveDeviceTokenToSNS = async(deviceToken: string, fetchClient: HttpClient): Promise<void> => {
   try {
     const response = await fetchClient.post('/notification/register', {
       deviceToken,
@@ -11,16 +9,9 @@ export const saveDeviceTokenToSNS = async(
     })
 
     if (response.status !== 200) {
-      const statusText = typeof response.statusText === 'string' ? response.statusText : 'Unknown error'
-      throw new Error(
-        `Failed to register device. Server responded with status: ${response.status} ${statusText}`
-      )
+      throw new Error(`Failed to register device. Server responded with status: ${response.status}`)
     }
   } catch (error) {
-    throw new Error(
-      `Failed to register device token. ${
-        error instanceof Error ? error.message : 'An unexpected error occurred'
-      }`
-    )
+    throw new Error(`Failed to register device token. ${error instanceof Error ? error.message : 'An unexpected error occurred'}`)
   }
 }
